endpoint-trips/requirements.txt
flask==0.12.2
python-google-places==1.4.1
python-dotenv==0.7.1